Improvements/Finalized CouchDB and PSQL seed scripts
