--- database/couchbase.config.js
@@ -1,7 +1,0 @@
-let credentials = {
-  username: 'Administrator',
-  password: 'password',
-  bucket: 'test'
-};
-
-module.exports = credentials;
