+++ database/seedpsql.js
@@ -15,13 +15,14 @@ const lorem = new LoremIpsum({
 
 let sectionCount = 1;
 let elementCount = 1;
+let courseCount = 1;
 let options = ['lecture', 'lecture', 'lecture', 'lecture', 'quiz', 'quiz', 'exercise', 'quiz', 'exercise', 'exercise', 'exercise', 'lecture', 'lecture', 'lecture', 'lecture', 'lecture', 'lecture', 'lecture', 'lecture', 'lecture', 'lecture', 'lecture', 'lecture', 'article', 'article', 'article'];
 
 
-const generateCourse = async (courseId) => {
+const generateCourse = async () => {
   const course = {
-    courseId: courseId,
-    totalSections: faker.datatype.number(30),
+    courseId: courseCount,
+    totalSections: faker.datatype.number(15),
     totalLectures: 0,
     totalExercises: 0,
     totalArticles: 0,
@@ -32,9 +33,9 @@ const generateCourse = async (courseId) => {
   return course;
 };
 
-const generateSection = (courseId, sectionCount, j) => {
+const generateSection = (j) => {
   const section = {
-    courseId: courseId, //added by jake
+    courseId: courseCount, //added by jake
     sectionId: sectionCount,
     title: faker.lorem.words(),
     sectionLength: 0,
@@ -119,4 +120,59 @@ const generate1mCourses = async (numberOfCourses) => {
   }
 };
 
-generate1mCourses(1000000);
+const generate5000Courses = async () => {
+  let courses = [];
+  let sections = [];
+  let elements = [];
+  for (let i = 1; i <= 5000 + 1; i++) {
+    let course = await generateCourse(i);
+    courseCount ++;
+    course.courseLength = new Date(course.courseLength);
+    for (let j = 1; j <= course.totalSections; j++) {
+      let section = await generateSection(i, sectionCount, j);
+      section.sectionLength = new Date(section.sectionLength);
+      sectionCount ++;
+      for (let k = 1; k <= faker.datatype.number(10); k++) {
+        let element = await generateElement(section.sectionId, k);
+        elementCount ++;
+        if (element.kind === 'lecture') {
+          section.lectures ++;
+          course.totalLectures ++;
+          section.sectionLength += element.elementLength.getTime();
+        }
+        if (element.kind === 'quiz') {
+          section.quizzes ++;
+          course.totalQuizzes ++;
+        }
+        if (element.kind === 'exercise') {
+          section.exercises ++;
+          course.totalExercises ++;
+        }
+        if (element.kind === 'article') {
+          section.articles ++;
+          course.totalArticles ++;
+          section.sectionLength += element.elementLength.getTime();
+          course.courseLength += element.elementLength.getTime();
+        }
+        elements.push(element);
+        // console.log('element saving', element.elementId);
+      }
+      sections.push(section);
+      // console.log('section saved', section.sectionId);
+    }
+    courses.push(course);
+    console.log('course created', course.courseId);
+  }
+  await pg.bulkCreateCourses(courses);
+  await pg.bulkCreateSections(sections);
+  await pg.bulkCreateElements(elements);
+  console.log(courseCount, 'courses saved to db');
+};
+
+const seedDB = async () => {
+  for (let s = 0; s <= 2000; s++) {
+    await generate5000Courses();
+  }
+};
+
+seedDB();
