+++ package.json
@@ -10,9 +10,9 @@
     "build": "webpack --mode production",
     "server-dev": "nodemon server/start.js",
     "seed": "node seedDatabase/index.js",
-    "seedPGSQL": "node database/seedpsql.js",
+    "seedPSQL": "node database/seedpsql.js",
     "syncPG": "node database/postgreSQL.js",
-    "seedCouch": "node database/CouchDB.js"
+    "seedCouch": "node database/CouchDBSeed.js"
   },
   "repository": {
     "type": "git",
@@ -32,7 +32,9 @@
     "compression": "^1.7.4",
     "compression-webpack-plugin": "^8.0.0",
     "cors": "^2.8.5",
+    "couch-db": "^1.1.3",
     "couchbase": "^3.1.3",
+    "couchdb": "^0.3.8",
     "date-fns": "^2.21.1",
     "download": "^8.0.0",
     "download-file": "^0.1.5",
@@ -47,6 +49,7 @@
     "moment-duration-format": "^2.3.2",
     "mongoose": "^5.12.9",
     "mp4-stream": "^3.1.3",
+    "nano": "^9.0.3",
     "nodemon": "^2.0.7",
     "path": "^0.12.7",
     "pexels": "^1.2.1",
